ocr/app/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pytesseract==0.3.10
Pillow==10.4.0
httpx==0.27.2
